Move stringify to its own files.
Replace TypeId diagnostic arguments with InstIdAs* where possible.
