+++ toolchain/check/handle_name.cpp
@@ -45,6 +45,8 @@ auto HandleParseNode(Context& context, Parse::PointerMemberAccessExprId node_id)
     -> bool {
   auto diagnose_not_pointer = [&context,
                                &node_id](SemIR::TypeId not_pointer_type_id) {
+    // TODO: Pass in the expression we're trying to dereference to produce a
+    // better diagnostic.
     CARBON_DIAGNOSTIC(ArrowOperatorOfNonPointer, Error,
                       "cannot apply `->` operator to non-pointer type `{0}`",
                       SemIR::TypeId);
