+++ toolchain/check/handle_operator.cpp
@@ -304,6 +304,8 @@ auto HandleParseNode(Context& context, Parse::PrefixOperatorStarId node_id)
   auto deref_base_id = PerformPointerDereference(
       context, node_id, base_id,
       [&context, &node_id](SemIR::TypeId not_pointer_type_id) {
+        // TODO: Pass in the expression we're trying to dereference to produce a
+        // better diagnostic.
         CARBON_DIAGNOSTIC(
             DerefOfNonPointer, Error,
             "cannot dereference operand of non-pointer type `{0}`",
