+++ toolchain/check/member_access.cpp
@@ -230,6 +230,7 @@ static auto PerformImplLookup(
       context, type_const_id, assoc_type.interface_type_id.AsConstantId());
   if (!witness_id.is_valid()) {
     if (missing_impl_diagnoser) {
+      // TODO: Pass in the expression whose type we are printing.
       CARBON_DIAGNOSTIC(MissingImplInMemberAccessNote, Note,
                         "type `{1}` does not implement interface `{0}`",
                         SemIR::NameId, SemIR::TypeId);
@@ -238,6 +239,7 @@ static auto PerformImplLookup(
                 context.GetTypeIdForTypeConstant(type_const_id))
           .Emit();
     } else {
+      // TODO: Pass in the expression whose type we are printing.
       CARBON_DIAGNOSTIC(MissingImplInMemberAccess, Error,
                         "cannot access member of interface `{0}` in type `{1}` "
                         "that does not implement that interface",
@@ -396,17 +398,17 @@ static auto PerformInstanceBinding(Context& context, SemIR::LocId loc_id,
 // Validates that the index (required to be an IntLiteral) is valid within the
 // tuple size. Returns the index on success, or nullptr on failure.
 static auto ValidateTupleIndex(Context& context, SemIR::LocId loc_id,
-                               SemIR::Inst operand_inst,
+                               SemIR::InstId operand_inst_id,
                                SemIR::IntLiteral index_inst, int size)
     -> const llvm::APInt* {
   const auto& index_val = context.ints().Get(index_inst.int_id);
   if (index_val.uge(size)) {
     CARBON_DIAGNOSTIC(TupleIndexOutOfBounds, Error,
-                      "tuple element index `{0}` is past the end of type `{1}`",
-                      TypedInt, SemIR::TypeId);
+                      "tuple element index `{0}` is past the end of type {1}",
+                      TypedInt, InstIdAsTypeOfExpr);
     context.emitter().Emit(loc_id, TupleIndexOutOfBounds,
                            {.type = index_inst.type_id, .value = index_val},
-                           operand_inst.type_id());
+                           operand_inst_id);
     return nullptr;
   }
   return &index_val;
@@ -429,10 +431,10 @@ auto PerformMemberAccess(Context& context, SemIR::LocId loc_id,
   auto base_type_id = context.insts().Get(base_id).type_id();
   if (!context.TryToCompleteType(base_type_id, [&] {
         CARBON_DIAGNOSTIC(IncompleteTypeInMemberAccess, Error,
-                          "member access into object of incomplete type `{0}`",
-                          SemIR::TypeId);
+                          "member access into object of incomplete type {0}",
+                          InstIdAsTypeOfExpr);
         return context.emitter().Build(base_id, IncompleteTypeInMemberAccess,
-                                       base_type_id);
+                                       base_id);
       })) {
     return SemIR::InstId::BuiltinError;
   }
@@ -462,18 +464,18 @@ auto PerformMemberAccess(Context& context, SemIR::LocId loc_id,
         }
       }
       CARBON_DIAGNOSTIC(QualifiedExprNameNotFound, Error,
-                        "type `{0}` does not have a member `{1}`",
-                        SemIR::TypeId, SemIR::NameId);
-      context.emitter().Emit(loc_id, QualifiedExprNameNotFound, base_type_id,
+                        "type {0} does not have a member `{1}`",
+                        InstIdAsTypeOfExpr, SemIR::NameId);
+      context.emitter().Emit(loc_id, QualifiedExprNameNotFound, base_id,
                              name_id);
       return SemIR::InstId::BuiltinError;
     }
 
     if (base_type_id != SemIR::TypeId::Error) {
       CARBON_DIAGNOSTIC(QualifiedExprUnsupported, Error,
-                        "type `{0}` does not support qualified expressions",
-                        SemIR::TypeId);
-      context.emitter().Emit(loc_id, QualifiedExprUnsupported, base_type_id);
+                        "type {0} does not support qualified expressions",
+                        InstIdAsTypeOfExpr);
+      context.emitter().Emit(loc_id, QualifiedExprUnsupported, base_id);
     }
     return SemIR::InstId::BuiltinError;
   }
@@ -520,11 +522,11 @@ auto PerformCompoundMemberAccess(
   // because the base expression is not used for anything.
   if (member_id == member_expr_id && member.type_id() != SemIR::TypeId::Error) {
     CARBON_DIAGNOSTIC(CompoundMemberAccessDoesNotUseBase, Error,
-                      "member name of type `{0}` in compound member access is "
+                      "member name of type {0} in compound member access is "
                       "not an instance member or an interface member",
-                      SemIR::TypeId);
+                      InstIdAsTypeOfExpr);
     context.emitter().Emit(loc_id, CompoundMemberAccessDoesNotUseBase,
-                           member.type_id());
+                           member_id);
   }
 
   return member_id;
@@ -534,16 +536,15 @@ auto PerformTupleAccess(Context& context, SemIR::LocId loc_id,
                         SemIR::InstId tuple_inst_id,
                         SemIR::InstId index_inst_id) -> SemIR::InstId {
   tuple_inst_id = ConvertToValueOrRefExpr(context, tuple_inst_id);
-  auto tuple_inst = context.insts().Get(tuple_inst_id);
-  auto tuple_type_id = tuple_inst.type_id();
+  auto tuple_type_id = context.insts().Get(tuple_inst_id).type_id();
 
   auto tuple_type = context.types().TryGetAs<SemIR::TupleType>(tuple_type_id);
   if (!tuple_type) {
     CARBON_DIAGNOSTIC(TupleIndexOnANonTupleType, Error,
-                      "type `{0}` does not support tuple indexing; only "
+                      "type {0} does not support tuple indexing; only "
                       "tuples can be indexed that way",
-                      SemIR::TypeId);
-    context.emitter().Emit(loc_id, TupleIndexOnANonTupleType, tuple_type_id);
+                      InstIdAsTypeOfExpr);
+    context.emitter().Emit(loc_id, TupleIndexOnANonTupleType, tuple_inst_id);
     return SemIR::InstId::BuiltinError;
   }
 
@@ -566,7 +567,7 @@ auto PerformTupleAccess(Context& context, SemIR::LocId loc_id,
   auto index_literal = context.insts().GetAs<SemIR::IntLiteral>(
       context.constant_values().GetInstId(index_const_id));
   auto type_block = context.type_blocks().Get(tuple_type->elements_id);
-  const auto* index_val = ValidateTupleIndex(context, loc_id, tuple_inst,
+  const auto* index_val = ValidateTupleIndex(context, loc_id, tuple_inst_id,
                                              index_literal, type_block.size());
   if (!index_val) {
     return SemIR::InstId::BuiltinError;
