+++ toolchain/check/sem_ir_diagnostic_converter.cpp
@@ -4,6 +4,7 @@
 
 #include "toolchain/check/sem_ir_diagnostic_converter.h"
 
+#include "toolchain/sem_ir/stringify_type.h"
 namespace Carbon::Check {
 
 auto SemIRDiagnosticConverter::ConvertLoc(SemIRLoc loc,
@@ -135,14 +136,17 @@ auto SemIRDiagnosticConverter::ConvertArg(llvm::Any arg) const -> llvm::Any {
   if (auto* type_id = llvm::any_cast<SemIR::TypeId>(&arg)) {
     // TODO: Format the enclosing "`"s here to prepare for adding an "aka" when
     // desirable.
-    return sem_ir_->StringifyType(*type_id);
+    return StringifyTypeExpr(*sem_ir_, sem_ir_->types().GetInstId(*type_id));
   }
   if (auto* typed_int = llvm::any_cast<TypedInt>(&arg)) {
     return llvm::APSInt(typed_int->value,
                         !sem_ir_->types().IsSignedInt(typed_int->type));
   }
   if (auto* type_expr = llvm::any_cast<InstIdAsType>(&arg)) {
-    return "`" + sem_ir_->StringifyTypeExpr(type_expr->inst_id) + "`";
+    return "`" + StringifyTypeExpr(*sem_ir_, type_expr->inst_id) + "`";
+  }
+  if (auto* type_expr = llvm::any_cast<InstIdAsRawType>(&arg)) {
+    return StringifyTypeExpr(*sem_ir_, type_expr->inst_id);
   }
   if (auto* type_of_expr = llvm::any_cast<InstIdAsTypeOfExpr>(&arg)) {
     if (!type_of_expr->inst_id.is_valid()) {
@@ -151,8 +155,10 @@ auto SemIRDiagnosticConverter::ConvertArg(llvm::Any arg) const -> llvm::Any {
     // TODO: Where possible, produce a better description of the type based on
     // the expression.
     return "`" +
-           sem_ir_->StringifyType(
-               sem_ir_->insts().Get(type_of_expr->inst_id).type_id()) +
+           StringifyTypeExpr(
+               *sem_ir_,
+               sem_ir_->types().GetInstId(
+                   sem_ir_->insts().Get(type_of_expr->inst_id).type_id())) +
            "`";
   }
   return DiagnosticConverter<SemIRLoc>::ConvertArg(arg);
